requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6; platform_system == "Windows"
exceptiongroup==1.3.0; python_version < "3.11"
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
huggingface-hub==0.31.2
idna==3.10
jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
libclang==18.1.1
linecache2==1.0.0
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.8.3.14; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-cupti-cu12==12.8.57; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-nvrtc-cu12==12.8.61; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-runtime-cu12==12.8.57; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cudnn-cu12==9.7.1.26; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufft-cu12==11.3.3.41; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufile-cu12==1.13.0.11; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-curand-cu12==10.3.9.55; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusolver-cu12==11.7.2.55; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparse-cu12==12.5.7.53; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparselt-cu12==0.6.3; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nccl-cu12==2.26.2; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvjitlink-cu12==12.8.61; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvtx-cu12==12.8.55; platform_system == "Linux" and platform_machine == "x86_64"
opt-einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.4
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1; python_version < "3.12"
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traceback2==1.4.0
transformers==4.51.3
triton==3.3.0; platform_system == "Linux"
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.2
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
--extra-index-url https://download.pytorch.org/whl/cu128
